refactor(core): Decouple server started event from internal hooks (no-changelog)
